Move time signature to new namespace directory
