+++ GDAPI/GDAPI/Utilities/Objects/Presets/GuidelineEditorPreset.cs
@@ -1,5 +1,6 @@
 ﻿using GDAPI.Utilities.Functions.Extensions;
 using GDAPI.Utilities.Objects.General;
+using GDAPI.Utilities.Objects.General.Music;
 using GDAPI.Utilities.Objects.GeometryDash.General;
 using System.Collections.Generic;
 using System.Text;
